Consolidate config loading with defaults
