+++ infrastructure/lib/types.ts
@@ -68,68 +68,3 @@ export type Config = {
 	pipeline: PipelineConfig;
 	app: AppConfig;
 };
-
-export const defaultConfig: Config = {
-	common: {
-		development: {
-			enable: false,
-		},
-		instance: {
-			name: "main",
-		},
-	},
-	pipeline: {
-		approvals: {
-			preCdkSynth: {
-				enable: true,
-				email: "",
-			},
-		},
-		removalPolicy: "retain",
-		source: {
-			repoBranch: "main",
-			repoHook: {
-				enable: false,
-			},
-			repoName: "document-translation",
-			repoOwner: "aws-samples",
-		},
-	},
-	app: {
-		cognito: {
-			localUsers: {
-				enable: false,
-				mfa: {
-					enforcement: "off",
-					otp: false,
-					sms: false,
-				},
-			},
-			saml: {
-				enable: false,
-				metadataUrl: "",
-			},
-		},
-		readable: {
-			enable: false,
-			bedrockRegion: "us-east-1",
-		},
-		removalPolicy: "retain",
-		translation: {
-			enable: false,
-			lifecycle: 7,
-			pii: {
-				enable: false,
-				lifecycle: 3,
-			},
-		},
-		webUi: {
-			customDomain: {
-				enable: false,
-				domain: "",
-				certificateArn: "",
-			},
-			enable: true,
-		},
-	},
-};
